Change int => u16 type for constants

+++ 2048-MotoMAGX/2048-MotoMAGX.cpp
@@ -32,14 +32,14 @@ struct Tile;
 typedef unsigned short int u16;
 typedef std::vector<Tile *> TileList;
 
-const int HORIZONTAL = 4, VERTICAL = 4;
-const int BOARD_SIZE = HORIZONTAL * VERTICAL;
-const int END_GAME_TARGET = 2048;
-const int TILE_SIZE = 32;
+const u16 HORIZONTAL = 4, VERTICAL = 4;
+const u16 BOARD_SIZE = HORIZONTAL * VERTICAL;
+const u16 END_GAME_TARGET = 2048;
+const u16 TILE_SIZE = 32;
 #if defined(EZX_EM30) || defined (EZX_E8)
-const int TILE_MARGIN = 5;
+const u16 TILE_MARGIN = 5;
 #else
-const int TILE_MARGIN = 16;
+const u16 TILE_MARGIN = 16;
 #endif
 
 inline double DegreesToRadians(int angleDegrees) { return ((angleDegrees) * M_PI / 180.0); }
